+++ src/application/validators/index.ts
@@ -1,3 +1,4 @@
 export * from './required-field'
 export * from './validator'
 export * from './composite'
+export * from './builder'
